+++ ChangeLog.md
@@ -8,6 +8,7 @@
 - simplify setup script: no longer installs sudo and util-linux
 - also setup home without runuser
 - add `--user` option (needed for ubuntu user)
+- `run --debug` now also outputs podman command like `--dryrun`
 
 ## 0.4.1 (2026-08-07)
 - require volume host paths to exist